[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BPF_LSM:

 Enables instrumentation of the security hooks with eBPF programs for
 implementing dynamic MAC and Audit Policies.

 If you are unsure how to answer this question, answer N.

 Symbol: BPF_LSM [=n]
 Type  : bool
 Defined at init/Kconfig:1623
   Prompt: LSM Instrumentation with BPF
   Depends on: BPF_EVENTS [=y] && BPF_SYSCALL [=y] && SECURITY [=y] && BPF_JIT [=y]
   Location:
     -> General setup

---