[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SENSORS_AMD_ENERGY:

 If you say yes here you get support for core and package energy
 sensors, based on RAPL MSR for AMD family 17h and above CPUs.

 This driver can also be built as a module. If so, the module
 will be called as amd_energy.

 Symbol: SENSORS_AMD_ENERGY [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:327
   Prompt: AMD RAPL MSR based Energy driver
   Depends on: HWMON [=y] && X86 [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])

---

 CONFIG_SENSORS_MAX16601:

 If you say yes here you get hardware monitoring support for Maxim
 MAX16601.

 This driver can also be built as a module. If so, the module will
 be called max16601.

 Symbol: SENSORS_MAX16601 [=n]
 Type  : tristate
 Defined at drivers/hwmon/pmbus/Kconfig:149
   Prompt: Maxim MAX16601
   Depends on: HWMON [=y] && PMBUS [=m]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> PMBus support (PMBUS [=m])

---

Cc: Dean Nelson <dnelson@redhat.com>
Signed-off-by: CKI@GitLab <cki-project@redhat.com>