New configs in drivers/phy
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PHY_QCOM_IPQ806X_USB:

 This option enables support for the Synopsis PHYs present inside the
 Qualcomm USB3.0 DWC3 controller on ipq806x SoC. This driver supports
 both HS and SS PHY controllers.

 Symbol: PHY_QCOM_IPQ806X_USB [=n]
 Type  : tristate
 Defined at drivers/phy/qualcomm/Kconfig:108
   Prompt: Qualcomm IPQ806x DWC3 USB PHY driver
   Depends on: HAS_IOMEM [=y] && OF [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> PHY Subsystem
 Selects: GENERIC_PHY [=y]

Cc: John Linville <linville@redhat.com>