[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MODPROBE_PATH:

 When kernel code requests a module, it does so by calling
 the "modprobe" userspace utility. This option allows you to
 set the path where that binary is found. This can be changed
 at runtime via the sysctl file
 /proc/sys/kernel/modprobe. Setting this to the empty string
 removes the kernel's ability to request modules (but
 userspace can still load modules explicitly).

 Symbol: MODPROBE_PATH [=/sbin/modprobe]
 Type  : string
 Defined at init/Kconfig:2302
   Prompt: Path to modprobe binary
   Depends on: MODULES [=y]
   Location:
     -> Enable loadable module support (MODULES [=y])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>