Fix fedora-only config updates
After every scheduled upstream merge, a script is run to find new
configs, generate pending-* files for them and for ark, create
branches to push MRs.

However, there is a case where new Fedora configs are found but not
new ARK configs.  This case had issues.

When the merge is done, make dist-configs-commit is called for Fedora
and RHEL.  This results in a possible new commit if configs are found.

The command is run twice and creates the following matrix:

1. Both Fedora and RHEL have new configs.
2. Neither Fedora nor RHEL have new configs.
3. Fedora does not have new configs, but RHEL does.
4. Fedora has new configs, bug RHEL does not.

1 and 2 are common and 1-3 are handled correctly.

Scenario 4 is not.

When the configs are committed, the next script gen_config_patches.sh
looks for a special commit 'AUTOMATIC'.  It doesn't care if that commit
comes from Fedora or RHEL.

The gen_config_patches.sh can not handle pending-fedora changes, so
those changes are filtered in the script.

In scenario 4 that leaves an empty file to process, which the script
aborts on.  Further no print out was displayed which slowed down
debugging.

The are multiple ways to address this.  This patch does multiple things.

* Adds a print statement on the failure for future debugging.
* Exits on success if there are no changes to process after filtering.
* Reduces the risk of calling gen_config_patches.sh by only calling
  the script when RHEL adds new configs.

Tested scenario 4 with last week's os-build HEAD.

Signed-off-by: Don Zickus <dzickus@redhat.com>