[redhat] New configs in net/ipv6
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: IPV6_IOAM6_LWTUNNEL [=n]
 Type  : bool
 Defined at net/ipv6/Kconfig:331
   Prompt: IPv6: IOAM Pre-allocated Trace insertion support
   Depends on: NET [=y] && INET [=y] && IPV6 [=y]
   Location:
     -> Networking support (NET [=y])
       -> Networking options
         -> TCP/IP networking (INET [=y])
           -> The IPv6 protocol (IPV6 [=y])
 Selects: LWTUNNEL [=y]

---

Cc: Jiri Benc <jbenc@redhat.com>
Cc: Marcelo Leitner <mleitner@redhat.com>
Cc: Antoine Tenart <atenart@redhat.com>
Cc: Balazs Nemeth <bnemeth@redhat.com>
Cc: Davide Caratti <dcaratti@redhat.com>
Cc: Eric Garver <egarver@redhat.com>
Cc: Flavio Leitner <fbl@redhat.com>
Cc: Florian Westphal <fwestpha@redhat.com>
Cc: Guillaume Nault <gnault@redhat.com>
Cc: Hangbin Liu <haliu@redhat.com>
Cc: Ivan Vecera <ivecera@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Cc: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Cc: Paolo Abeni <pabeni@redhat.com>
Cc: Phil Sutter <psutter@redhat.com>
Cc: Sabrina Dubroca <sdubroca@redhat.com>
Cc: Stefano Brivio <sbrivio@redhat.com>
Cc: Xin Long <lxin@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>