redhat: move the DIST variable setting to Makefile.variables
The DIST value is dynamically detected on ARK, but it can be set to a
fixed value for a branch which forks from it (as some branches are built
only for a given distro and we don't want the DIST value to change even
if building eg. the src.rpm in a different environment). So it makes sense
in this case to move it to the Makefile.variables file.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>