redhat/genspec.sh: Rearrange genspec.sh
genspec.sh takes a long time to run when executed to create test
examples.  The script does three things: creates a changelog, creates a
patch file, and creates the specfile.  Reorganize the code so that the
specfile is created first, and exit if the file is created for a test
examples.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>