[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM64_SME [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:1952
   Prompt: ARM Scalable Matrix Extension support
   Depends on: ARM64_SVE [=y]
   Location:
     Main menu
       -> Kernel Features
         -> ARM Scalable Vector Extension support (ARM64_SVE [=y])

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>