[redhat] New configs in drivers/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: S390_UV_UAPI [=m]
 Type  : tristate
 Defined at drivers/s390/char/Kconfig:103
   Prompt: Ultravisor userspace API
   Depends on: S390 [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Character devices

---

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>

v2:
- CONFIG_S390_UV_UAPI enabled as an s390x builtin in ark.
- CONFIG_S390_UV_UAPI disabled for zfcpdump in ark.