[redhat] New configs in drivers/virt
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NITRO_ENCLAVES_MISC_DEV_TEST [=y]
 Type  : bool
 Defined at drivers/virt/nitro_enclaves/Kconfig:18
   Prompt: Tests for the misc device functionality of the Nitro Enclaves
   Depends on: VIRT_DRIVERS [=y] && NITRO_ENCLAVES [=m] && KUNIT [=m]
   Visible if: VIRT_DRIVERS [=y] && NITRO_ENCLAVES [=m] && KUNIT [=m] && !KUNIT_ALL_TESTS [=m]
   Location:
     Main menu
       -> Device Drivers
         -> Virtualization drivers (VIRT_DRIVERS [=y])
           -> Nitro Enclaves Support (NITRO_ENCLAVES [=m])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>