[redhat] New configs in drivers/gpu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DRM_RCAR_USE_MIPI_DSI [=n]
 Type  : bool
 Defined at drivers/gpu/drm/rcar-du/Kconfig:47
   Prompt: R-Car DU MIPI DSI Encoder Support
   Depends on: HAS_IOMEM [=y] && DRM_BRIDGE [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> R-Car DU MIPI DSI Encoder Support (DRM_RCAR_USE_MIPI_DSI [=n])

---

Cc: David Airlie <airlied@redhat.com>
Cc: Adam Jackson <ajax@redhat.com>
Cc: Lyude Paul <lyude@redhat.com>
Cc: "Michel Dänzer" <mdaenzer@redhat.com>
Cc: Karol Herbst <kherbst@redhat.com>
Cc: Ben Skeggs <bskeggs@redhat.com>
Cc: Jocelyn Falempe <jfalempe@redhat.com>
Cc: "Mika Penttilä" <mpenttil@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>