Trim changelog for 6.2 cycle
For some reason, when we move to a new kernel cycle, the changelog gits
confused, and we get every entry repeated on the first build.  To keep
the changelog from growing out of control with repeats, I trim it,
though it would be nice to keep original dates and solve this problem.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>