[redhat] New configs in drivers/vfio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: VFIO_AMBA [=n]
 Type  : tristate
 Defined at drivers/vfio/platform/Kconfig:19
   Prompt: VFIO support for AMBA devices
   Depends on: VFIO [=m] && (ARM || ARM64 [=y] || COMPILE_TEST [=n]) && (ARM_AMBA [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> VFIO Non-Privileged userspace driver framework (VFIO [=m])
         -> VFIO support for platform devices
           -> VFIO support for AMBA devices (VFIO_AMBA [=n])
 Selects: VFIO_PLATFORM_BASE [=n]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>