redhat/kernel.spec.template: update compression variables to support zstd
Upstream Status: RHEL Only

The kernel.spec.template file has a compression variable to define which
binary is used to compress the kernel modules. The binary is currently
invoked in the following manner: '%compression -c9'. There is a desire
to use zstd for automotive since it results in faster decompression,
and faster boot times. To support zstd, different flags need to be
passed to the compression binary, so let's introduce a compression_flags
variable.

The BuildRequires needs to be updated to include the compression
binary as well since these variables can be overridden on the command
line with "rpmbuild --define 'compression zstd' ...".

The zstd support was tested with the following configuration:

    %global compression zstd
    %global compression_flags --rm
    %global compext zst

Note that the --stdout flag that's passed to the compression binary
works with bz2, gz, xz, and zstd.

Signed-off-by: Brian Masney <bmasney@redhat.com>