New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MEMFD_CREATE [=n]
 Type  : bool
 Defined at mm/Kconfig:1166
   Prompt: Enable memfd_create() system call
   Visible if: EXPERT [=y]
   Location:
     -> Memory Management options
       -> Enable memfd_create() system call (MEMFD_CREATE [=n])
 Selected by [n]:
   - TMPFS [=n] && SHMEM [=y]
   - HUGETLBFS [=n] && (X86 || IA64 || SPARC64 || ARCH_SUPPORTS_HUGETLBFS [=y] || BROKEN [=n]) && (SYSFS [=y] || SYSCTL [=y])